Move files into HomeNav
